+++ packages/cheqd-blockchain-api/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @docknetwork/cheqd-blockchain-api
 
+## 4.0.7
+
+### Patch Changes
+
+- Delegation engine
+- Updated dependencies
+  - @docknetwork/credential-sdk@0.54.11
+
 ## 4.0.6
 
 ### Patch Changes